Move MinimumDataRate to Features namespace.
