+++ src/Microsoft.AspNetCore.Server.Kestrel.Core/Features/MinimumDataRate.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.AspNetCore.Server.Kestrel.Core
+namespace Microsoft.AspNetCore.Server.Kestrel.Core.Features
 {
     public class MinimumDataRate
     {
